requirements.txt
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.2.10
asn1crypto==0.24.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==5.0.1
blessed==1.17.11
blinker==1.4
blis==0.7.8
boto3==1.24.84
botocore==1.27.84
breadability==0.1.20
bs4==0.0.1
catalogue==2.0.8
certifi==2020.4.5.1
cffi==1.14.4
cfgv==3.2.0
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.3
confection==0.0.2
cryptography==2.8
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
Django==3.1.2
django-ckeditor==6.0.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-js-asset==1.2.2
django-picklefield==3.0.1
django-q==1.3.4
django-restframework==0.0.1
djangorestframework==3.12.1
djangorestframework-api-key==2.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docopt==0.4.0
en-core-web-sm==3.4.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fire==0.3.1
fundamentals==2.2.4
future==0.18.2
gensim==3.8.3
googletrans==2.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
httpcore==0.9.1
httplib2==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==1.5.9
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
jellyfish==0.7.2
Jinja2==2.11.3
jmespath==1.0.1
joblib==0.15.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
keyring==18.0.1
keyrings.alt==3.4.0
langcodes==3.3.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==4.6.1
macaroonbakery==1.3.1
mailchimp-transactional==1.0.21
Mako==1.1.0
mandrill==1.0.60
Markdown==3.1.1
MarkupSafe==1.1.0
mistune==0.8.4
monotonic==1.5
Morfessor==2.0.6
moviepy==1.0.3
murmurhash==1.0.2
natsort==8.2.0
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.4
nltk==3.5
nodeenv==1.5.0
notebook==6.2.0
numpy==1.18.4
oauth==1.0.1
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.5.2
pathy==0.6.2
pbr==5.5.1
pdfminer==20191125
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
pluggy==0.13.1
polyglot==16.7.4
polygot==0.1
pre-commit==2.8.2
preshed==3.0.2
profanity-check==1.0.3
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.7.3
ptyprocess==0.7.0
py==1.9.0
pycld2==0.41
pycodestyle==2.6.0
pycountry==19.8.18
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.9
pydantic==1.9.2
pydub==0.24.1
Pygments==2.7.4
PyICU==2.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyMuPDF==1.18.8
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.14.1
PyQt5-sip==12.11.0
pyRFC3339==1.1
pyrsistent==0.17.3
pysbd==0.2.3
pytest==6.1.2
python-dateutil==2.8.1
python-docx==0.8.10
python-magic==0.4.27
python-pptx==0.6.18
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
rake-nltk==1.0.4
readability-api==1.0.2
redis==3.5.3
regex==2020.5.14
reportlab==3.5.34
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
rmtree==1.0
rollbar==0.15.1
s3transfer==0.6.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.4.1
SecretStorage==2.3.1
segtok==1.5.10
Send2Trash==1.5.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
simplejson==3.16.0
six==1.15.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SpeechRecognition==3.8.1
sqlparse==0.4.1
srsly==2.4.4
stdeb==0.8.5
sumy==0.8.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
testresources==2.0.1
thinc==8.1.2
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.7.1
tornado==6.1
tox==3.20.1
tqdm==4.46.0
traitlets==4.3.3
transformers==4.3.2
translate==3.5.0
typer==0.4.2
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.11
virtualenv==20.1.0
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
whitenoise==5.2.0
XlsxWriter==1.3.7
yake==0.3.7
zip-files==0.3.0
zope.interface==4.7.1
